+++ tests/Tests.hs
@@ -8,6 +8,7 @@ import Test.Framework (defaultMain)
 
 import qualified Tests.Properties as Properties
 import qualified Tests.Regressions as Regressions
+import qualified Tests.Unit as Unit
 
 main :: IO ()
-main = defaultMain [Properties.tests, Regressions.tests]
+main = defaultMain [Properties.tests, Regressions.tests, Unit.tests]